fix(agent): move ApiGatewayRequiredDialog from per-tab to window level
Move the API gateway required prompt dialog from AgentChat (per-tab) to
AppShell (window-level). Fixes a bug where the prompt would be missed if
the originating agent tab became dormant before the dialog could appear.

Also removes duplicate IPC subscriptions that arose when multiple agent
tabs were active simultaneously.
